Allow branch deletion when the final policy disables deletion.
The initial implementation allows branch deletion only when the
final policy in the policy array (the fallback policy) has approval
policy "off". This is to ensure that the fallback policy doesn't
delete branches willy-nilly.

It is equivalent to have the 'delete' flag disabled in the final
policy in the policy array. If we test the 'delete' flag then
we also allow for other policies in the final policy other than
the "off" policy.